requirements.txt
asyncpg~=0.29.0
python-dotenv~=1.0.1
aiogram~=3.13.1